pyproject.toml
[tool.poetry]
name = "RadarTimelineData"
version = "0.1.0"
description = ""
authors = ["oliver.reeves <oliver.reeves@ukkidney.org>"]
readme = "README.md"
packages = [{include = "radar_timeline_data"}]

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0"
sqlalchemy = "^2.0.23"
pyodbc = ">=5.0.1,<6.0.0"
xlsxwriter = "^3.1.0"
openpyxl = "^3.1.2"
psycopg2 = "^2.9.6"
types-python-dateutil = "^2.8.19.14"
types-openpyxl = "^3.1.0.26"
polars = "0.20.13"
tox = "^4.14.1"
pandas = "1.5.3"
rr-connection-manager = "^0.2.0"
python-docx = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.3.1"
pytest-mock = "^3.10.0"
faker = "^18.4.0"
mypy = "^1.2.0"
bandit = "^1.7.5"
tox = "^4.4.12"
pytest-cov = "^4.0.0"
pylint = "^3.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint]
disable = "too-many-locals, too-many-return-statements, too-many-branches, fixme, too-many-lines"
max-line-length = 160